feat: add shadcn plant form component
